requirements_dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements_dev.txt requirements_dev.in
#
appdirs==1.4.4            # via black
argparse==1.4.0           # via -r ./requirements.txt, dateutils
asgiref==3.2.10           # via -r ./requirements.txt, django
attrs==19.3.0             # via black, pytest
backcall==0.2.0           # via ipython
black==19.10b0            # via -r requirements_dev.in, pytest-black
certifi==2020.6.20        # via -r ./requirements.txt, requests, sentry-sdk
cffi==1.14.2              # via -r ./requirements.txt, cryptography
chardet==3.0.4            # via -r ./requirements.txt, requests
click==7.1.2              # via black, pip-tools
coreapi==2.3.3            # via -r ./requirements.txt, drf-yasg
coreschema==0.0.4         # via -r ./requirements.txt, coreapi, drf-yasg
coverage==5.2.1           # via pytest-cov
cryptography==3.0         # via -r ./requirements.txt, jwcrypto
datapunt-authorization-django==1.0.0  # via -r ./requirements.txt
datapunt-authorization-levels==0.1.3  # via -r ./requirements.txt
dateutils==0.6.8          # via -r ./requirements.txt
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via -r ./requirements.txt, djangorestframework-xml
django-extensions==3.0.5  # via -r ./requirements.txt
django-filter==2.3.0      # via -r ./requirements.txt
django==3.0               # via -r ./requirements.txt, datapunt-authorization-django, django-filter, djangorestframework, drf-hal-json, drf-nested-fields, drf-yasg
djangorestframework-csv==2.1.0  # via -r ./requirements.txt, drf-amsterdam
djangorestframework-gis==0.15  # via -r ./requirements.txt
djangorestframework-xml==2.0.0  # via -r ./requirements.txt, drf-amsterdam
djangorestframework==3.11.1  # via -r ./requirements.txt, djangorestframework-csv, djangorestframework-gis, drf-amsterdam, drf-extensions, drf-hal-json, drf-nested-fields, drf-yasg
drf-amsterdam==0.3.2      # via -r ./requirements.txt
drf-extensions==0.6.0     # via -r ./requirements.txt, drf-amsterdam
drf-hal-json==0.10.0      # via -r ./requirements.txt
drf-nested-fields==0.9.5  # via -r ./requirements.txt, drf-hal-json
drf-yasg==1.17.1          # via -r ./requirements.txt
et-xmlfile==1.0.1         # via -r ./requirements.txt, openpyxl
factory-boy==3.0.1        # via -r requirements_dev.in
faker==4.1.1              # via factory-boy
flake8==3.8.3             # via pytest-flake8
idna==2.10                # via -r ./requirements.txt, requests
inflection==0.5.0         # via -r ./requirements.txt, drf-yasg
iniconfig==1.0.1          # via pytest
ipdb==0.13.3              # via -r requirements_dev.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.17.0           # via ipdb
isort==5.4.2              # via pytest-isort
itypes==1.2.0             # via -r ./requirements.txt, coreapi
jdcal==1.4.1              # via -r ./requirements.txt, openpyxl
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via -r ./requirements.txt, coreschema
jwcrypto==0.7             # via -r ./requirements.txt, datapunt-authorization-django
markupsafe==1.1.1         # via -r ./requirements.txt, jinja2
mccabe==0.6.1             # via flake8
more-itertools==8.4.0     # via pytest
openpyxl==3.0.4           # via -r ./requirements.txt
packaging==20.4           # via -r ./requirements.txt, drf-yasg, pytest
parso==0.7.1              # via jedi
pathspec==0.8.0           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.3.1          # via -r requirements_dev.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.6     # via ipython
psycopg2-binary==2.8.5    # via -r ./requirements.txt
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8
pycparser==2.20           # via -r ./requirements.txt, cffi
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via ipython
pyparsing==2.4.7          # via -r ./requirements.txt, packaging
pytest-black==0.3.10      # via -r requirements_dev.in
pytest-cov==2.10.1        # via -r requirements_dev.in
pytest-django==3.9.0      # via -r requirements_dev.in
pytest-flake8==1.0.6      # via -r requirements_dev.in
pytest-isort==1.1.0       # via -r requirements_dev.in
pytest==6.0.1             # via -r requirements_dev.in, pytest-black, pytest-cov, pytest-django, pytest-flake8, pytest-isort
python-dateutil==2.8.1    # via -r ./requirements.txt, dateutils, faker
python-swiftclient==3.10.0  # via -r ./requirements.txt
pytz==2020.1              # via -r ./requirements.txt, dateutils, django
regex==2020.7.14          # via black
requests==2.24.0          # via -r ./requirements.txt, coreapi, datapunt-authorization-django, python-swiftclient
ruamel.yaml.clib==0.2.0   # via -r ./requirements.txt, ruamel.yaml
ruamel.yaml==0.16.10      # via -r ./requirements.txt, drf-yasg
sentry-sdk==0.16.5        # via -r ./requirements.txt
six==1.15.0               # via -r ./requirements.txt, cryptography, djangorestframework-csv, drf-yasg, packaging, pip-tools, python-dateutil, python-swiftclient, traitlets
sqlparse==0.3.1           # via -r ./requirements.txt, django
text-unidecode==1.3       # via faker
toml==0.10.1              # via black, pytest, pytest-black
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black
unicodecsv==0.14.1        # via -r ./requirements.txt, djangorestframework-csv
uritemplate==3.0.1        # via -r ./requirements.txt, coreapi, drf-yasg
urllib3==1.25.10          # via -r ./requirements.txt, requests, sentry-sdk
wcwidth==0.2.5            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
